refactor(cli): Add Publisher helper to encapsulate publishing logic
Extracted publishing operations into a dedicated Publisher struct to reduce parameter passing and improve code organization.
